Fixed Long Map ID Issues

+++ DeluxeMediaPlugin/src/main/java/com/github/pulsebeat02/deluxemediaplugin/command/ImageCommand.java
@@ -126,7 +126,7 @@ public class ImageCommand extends AbstractCommand implements CommandExecutor, Li
               break;
             }
           }
-          MapImage.resetMap(getPlugin().getLibrary(), id);
+          MapImage.resetMap(getPlugin().getLibrary(), (int)id);
           sender.sendMessage(
               ChatUtilities.formatMessage(
                   ChatColor.GOLD + "Successfully purged the map with ID " + id));
@@ -165,17 +165,19 @@ public class ImageCommand extends AbstractCommand implements CommandExecutor, Li
             return;
           }
           final String mrl = args[3];
+          // if it's between the bounds, it is safe to cast to integer because the original value is
+          // still being stored, and there is no integer overflow/
           if ((mrl.startsWith("http://")) || mrl.startsWith("https://") && mrl.endsWith(".png")) {
             final File img =
                 FileUtilities.downloadImageFile(mrl, getPlugin().getDataFolder().getAbsolutePath());
-            new MapImage(getPlugin().getLibrary(), id, img, width, height).drawImage();
+            new MapImage(getPlugin().getLibrary(), (int)id, img, width, height).drawImage();
             sender.sendMessage(
                 ChatUtilities.formatMessage(
                     ChatColor.GOLD + "Successfully drew image on map " + id));
           } else {
             final File f = new File(getPlugin().getDataFolder(), mrl);
             if (f.exists()) {
-              new MapImage(getPlugin().getLibrary(), id, f, width, height).drawImage();
+              new MapImage(getPlugin().getLibrary(), (int)id, f, width, height).drawImage();
               sender.sendMessage(
                   ChatUtilities.formatMessage(
                       ChatColor.GOLD + "Successfully drew image on map " + id));
